+++ routes/builderRoutes/search.js
@@ -2,12 +2,43 @@ const config = require('../../real_config.js')
 const firebase = require('../../firebase.js').builder
 const Fuse = require('fuse.js')
 
-// Import media types
+// Import content types
 const SeriesReal = require('../../models/builderModels/models-real/Series.js')
 const SeriesOther = require('../../models/builderModels/models-other/Series.js')
 const LessonOther = require('../../models/builderModels/models-other/Lesson.js')
 
 const Topic = require('../../models/nqModels/Topic.js')
+
+// Import NQ media types
+const Book = require('../../models/nqModels/Book.js')
+const Movie = require('../../models/nqModels/Movie.js')
+const Image = require('../../models/nqModels/Image.js')
+const Video = require('../../models/nqModels/Video.js')
+const Article = require('../../models/nqModels/Article.js')
+const Note = require('../../models/nqModels/Note.js')
+const Doc = require('../../models/nqModels/Document.js')
+const Discourse = require('../../models/nqModels/Discourse.js')
+const Composition = require('../../models/nqModels/Composition.js')
+const Quote = require('../../models/nqModels/Quote.js')
+const Outline = require('../../models/nqModels/Outline.js')
+const Idea = require('../../models/nqModels/Idea.js')
+const Illustration = require('../../models/nqModels/Illustration.js')
+
+const nqMedia = {
+    'book': Book,
+    'movie': Movie,
+    'image': Image,
+    'video': Video,
+    'article': Article,
+    'note': Note,
+    'document': Doc,
+    'discourse': Discourse,
+    'composition': Composition,
+    'quote': Quote,
+    'outline': Outline,
+    'idea': Idea,
+    'illustration': Illustration
+}
 
 // Import other media types
 const OQuote = require('../../models/builderModels/models-other/Quote.js')
@@ -34,6 +65,7 @@ var otherContent = {
 }
 
 module.exports = function (req, res) {
+    console.log('--builder search run--')
     console.log('type', req.body.type)
     console.log('search', req.body.terms)
     var token = req.body.token
@@ -139,8 +171,44 @@ module.exports = function (req, res) {
                     console.log('error with search media', err)
                     res.status(400).send('Bad search')
                 })
+            } else if (type === 'nqmedia') {
+                console.log('nq media search')
+                Promise.all(Object.keys(nqMedia).map((e) => { return nqMedia[e].find().lean().exec()})).then((items) => {
+                    // console.log('init items', items)
+                    var listTypes = Object.keys(nqMedia)
+                    var allItems = []
+                    items.forEach((listType, index) => {
+                        listType.forEach((item) => {
+                            item.type = listTypes[index]
+                            console.log('item', item, listTypes[index])
+                            allItems.push(item)
+                        })
+                    })
+                    // var allItems = items[0].concat(items[1], items[2], items[3], items[4])
+                    // Search with fuse
+                    var options = {
+                        keys: [{
+                            name: 'text',
+                            weight: 0.1
+                        }, {
+                            name: 'title',
+                            weight: 0.3
+                        }, {
+                            name: 'tags',
+                            weight: 0.6
+                        }]
+                    }
+                    var search = new Fuse(allItems, options)
+                    var result = search.search(terms)
+                    console.log(result)
+                    res.send(result)
+                })
+                .catch(function(err) {
+                    console.log('error with search media', err)
+                    res.status(400).send('Bad search')
+                })
             } else {
-                console.log('something went wrong')
+                console.log('something went wrong - search')
             }
         })
         .catch(function(err) {